requirements.txt
paho-mqtt
local_resources
